moved the effective 1RM methods (within the Effective1RMUtils class) into its own utils file within constants, out of the general_set_container_widget.dart file, also moved the enums into the enum file
